[fix] TripControllerTest 통합 및 에러 수정 완료
